requirements.txt
discord.py
discord.py[voice]
python-dotenv
youtube_dl